ImageDetectAPI-Kubernate/backend/requirements.txt
Flask
pillow
numpy
opencv-python
